refactor: move dist utils to separate file, unifi logging.
